Use float as cell type for new tensor-valued features
The two recently added tensor features, queryTermDocumentFrequency(field)
and bm25_for_labels(field), built their output with double cells, but
the values - bm25 scores and document frequencies - do not need double
precision, and keeping the cells in float lets downstream expressions
stay in float (for performance) instead of widening the whole computation.

The declared cell type has to agree in two places: the backend blueprints in
searchlib, and the vespa.type.feature property that the config model derives
for the ranking expression type checker (MapEvaluationTypeContext). Both are
changed here, and the unit tests are updated to track the change.

Also adds queryTermDocumentFrequency to the summary-features of the
with_corpus_features test profile, so the derived vespa.type.feature
property for it is actually covered by the rankingexpression
derived-config test.

+++ config-model/src/main/java/com/yahoo/schema/MapEvaluationTypeContext.java
@@ -312,12 +312,12 @@ public class MapEvaluationTypeContext extends FunctionReferenceContext implement
         if (reference.name().equals("queryTermDocumentFrequency")) {
             if (reference.arguments().size() != 1)
                 throw new IllegalArgumentException(reference.name() + " must have one argument");
-            return Optional.of(new TensorType.Builder(TensorType.Value.DOUBLE).mapped("term").build());
+            return Optional.of(new TensorType.Builder(TensorType.Value.FLOAT).mapped("term").build());
         }
         if (reference.name().equals("bm25_for_labels")) {
             if (reference.arguments().size() != 1)
                 throw new IllegalArgumentException(reference.name() + " must have one argument");
-            return Optional.of(new TensorType.Builder(TensorType.Value.DOUBLE).mapped("label").build());
+            return Optional.of(new TensorType.Builder(TensorType.Value.FLOAT).mapped("label").build());
         }
         if ( ! reference.name().equals("tensorFromLabels") &&
              ! reference.name().equals("tensorFromLabelsWithOffset") &&
